requirements.txt
ultralytics
uvicorn[standard]
fastapi
Pillow
opencv-python
